Revisions based on co-work and research
Documented the pagination method using Link headers.  Added a
reference to the blog post the API design is based on, with some
discussion.  Added additional indices that will be required for
history pagination.  Added documentation of API errors.  Removed
the links from tokens in favor of simple token identifiers.  Added
details on how to handle expiration of internal tokens and details
on how the caching works.